eval_framework/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "company-eval-framework"
version = "0.1.0"
description = "Reusable evaluation framework for LLM apps"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Company", email = "team@company.com"}
]

dependencies = [
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
company-eval = "company_eval_framework.cli:main"

[tool.setuptools]
packages = ["company_eval_framework"]
package-dir = {"" = "src"}
